requirements.txt
streamlit>=1.34.0
tensorflow~=2.16.0
opencv-python-headless
numpy
Pillow
